re-organized the code
